requirements.txt
openai 
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic==1.10.8
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff